requirements.txt
cfgv==3.4.0
distlib==0.3.8
filelock==3.15.4
identify==2.6.0
iniconfig==2.0.0
nodeenv==1.9.1
numpy==2.0.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.1
pip==24.2
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
pyserial==3.5
pytest==8.3.2
pytest-mock==3.14.0
pyyaml==6.0.2
virtualenv==20.26.3
